% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localFaov-package.R
\docType{package}
\name{localFaov-package}
\alias{localFaov}
\alias{localFaov-package}
\title{localFaov: Perform (local) functional ANOVA over a signal}
\description{
This package is designed to perform domain selection over a signal, i.e find portion of a signal responsible for the difference between curves.
}
\author{
\strong{Maintainer}: David Causeur \email{david.causeur@institut-agro.fr} (\href{https://orcid.org/0000-0001-6910-9440}{ORCID})

Authors:
\itemize{
  \item Rémi Mahmoud (\href{https://orcid.org/0000-0003-0853-0834}{ORCID})
}

Other contributors:
\itemize{
  \item Ching-Fan Sheu (\href{https://orcid.org/0000-0002-5978-5768}{ORCID}) [data contributor, contractor]
}

}
\keyword{internal}
